refactor: move integration tests and fix links and move jest files up by a folder
