+++ src/modules/users/tests/e2e/users.e2e-spec.ts
@@ -4,15 +4,12 @@ import { UsersRole } from "@prisma/client";
 import { HttpStatus } from "@nestjs/common";
 import * as pactum from "pactum";
 import { randomUUID } from "crypto";
-import { PrismaService } from "../../../src/config/prisma/prisma.service";
-import { AppModule } from "../../../src/app.module";
-import { SignInAuthDto } from "../../../src/modules/auth/dto";
-import {
-  CreateUserDto,
-  UpdateUserRoleDto
-} from "../../../src/modules/users/dto";
-import { UsersService } from "../../../src/modules/users/users.service";
-import { AuthService } from "../../../src/modules/auth/auth.service";
+import { PrismaService } from "../../../../config/prisma/prisma.service";
+import { AppModule } from "../../../../app.module";
+import { SignInAuthDto } from "../../../auth/dto";
+import { CreateUserDto, UpdateUserRoleDto } from "../../../users/dto";
+import { UsersService } from "../../users.service";
+import { AuthService } from "../../../auth/auth.service";
 
 describe("/users", () => {
   let app: INestApplication;
