+++ tests/jest-e2e.json
@@ -1,8 +1,8 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": "../../src/modules/",
+  "rootDir": "../src/modules/",
   "testEnvironment": "node",
-  "testRegex": ".unit-spec.ts$",
+  "testRegex": ".e2e-spec.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   }
