+++ tests/jest-int.json
@@ -1,8 +1,8 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": ".",
+  "rootDir": "../src/modules/",
   "testEnvironment": "node",
-  "testRegex": ".e2e-spec.ts$",
+  "testRegex": ".int-spec.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   }
