+++ tests/jest-unit.json
@@ -1,8 +1,8 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": ".",
+  "rootDir": "../src/modules/",
   "testEnvironment": "node",
-  "testRegex": ".int-spec.ts$",
+  "testRegex": ".unit-spec.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   }
